requirements.txt
pyrogram
TgCrypto
requests
python-dotenv
bs4
aiofiles
aiohttp